updated component and seeds

+++ client/src/Components/Game.js
@@ -1,32 +1,60 @@
 import React, {useState} from "react";
 import ReviewForm from "./ReviewForm";
-import Reviews from "./Reviews";
+import ReviewList from "./ReviewList";
 
 
-function Game (props, { onDeleteGame }) {
+function Game ({onDeleteGame, name, image_url, id, games, game, reviews, setReviews}) {
 
+  const [showForm, setShowForm] = useState(false);
+  
 
 
     function handleDeleteClick() {
-        fetch(`./games/${props.id}`, {
+      if (window.confirm("Are you sure you want to delete this game?")) {
+        fetch(`./games/${id}`, {
           method: "DELETE",
           headers: {
             "Content-Type": "application/json",
-         }})
+          },
+        })
           .then((r) => r.json())
-          .then(() => onDeleteGame(props));
+          .then(() => {
+            console.log("Deleting game:", id);
+            if (game) {
+              onDeleteGame(game);
+            }    
+          })}};
+
+      function handleAddReview(newReview){
+        setReviews([...reviews, newReview])
       }
 
+      function handleAddReviewClick() {
+        setShowForm(true);
+      }
+
+      function handleCancelAddReview() {
+    setShowForm(false);
+  }
 
     return (
         <div className="cards">
-            <h3>{props.name}</h3>
-            <img alt="No Image Available" src={props.image_url}></img>
+            <h3>{name}</h3>
+            <img className="image" alt="No Image Available" src={image_url}></img>
             <br></br>
-            <button className="delete" onClick={handleDeleteClick}>Delete Game</button>
-            <Reviews />
-            <ReviewForm />
+            {/* {console.log(id)} */}
+            <ReviewList game_id={id} />
+            <button className="button" onClick={handleAddReviewClick} onAddReview={handleAddReview}>Add Review</button>
+            {showForm && (
+              <>
+                <ReviewForm  onAddReview={handleAddReview} game_id={game?.id}/>
+                <button className="button" onClick={handleCancelAddReview}>Cancel</button>
+              </>
+            )}
+            <br></br>
+             <button className="delete" onClick={handleDeleteClick}>Delete Game</button>
         </div>
+        
     )
 
 }
